Fixes Bug in time.isBetweenTimes
The test in isBetweenTimes is backwards. If the endTime is before the start time, assume the time range spans midnight. In that case we need to flip the condition to test if the time is after the *end time*  or before the *start time*. I'm not sure how this got missed during testing but I just ran into it big time in one of my rule templates (it uncovered another error so it's all good :-) ).

Signed-off-by: Richard Koshak <rlkoshak@gmail.com>